Breaking up set of 5 .combine() tests
I'm not familiar with cats and I had trouble with the first question here.  When the answer I submitted failed, I could not tell whether it was because I didn't understand how .combine() worked at all, or because one of my answers was wrong.  Since the first 4 are roughy the same, I bumped the 5th into it's own test.